backend/requirements.txt
django==5.0.7
djangorestframework==3.16.0
mysqlclient==2.2.4
python-dotenv==1.0.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
openpyxl==3.0.7
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
